[FE #47] RENAME: 폴더, 경로변경
