+++ frontend/src/components/RoomList/RoomListItem.tsx
